app/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.3
psycopg[binary]==3.2.1
requests==2.32.3
tenacity==9.0.0
python-dotenv==1.0.1
whitenoise==6.7.0
dj-database-url==2.2.0
django-redis==5.4.0
gunicorn==22.0.0